Split code into modules
- start building DDD modules

+++ app/src/main/java/au/com/beba/runninggoal/feature/goals/WorkoutViewHolder.kt
@@ -5,7 +5,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import au.com.beba.runninggoal.R
 import au.com.beba.runninggoal.component.NumericWithLabel
-import au.com.beba.runninggoal.domain.core.Workout
+import au.com.beba.runninggoal.domain.Workout
 import au.com.beba.runninggoal.domain.core.Distance
 import au.com.beba.runninggoal.domain.core.GoalDate
 import org.jetbrains.anko.find
@@ -19,7 +19,7 @@ class WorkoutViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
     fun bindView(workout: Workout) {
         val dt = GoalDate(workout.dateTime)
-        val d = Distance.fromMetres(workout.distanceInMetres)
+        val d = Distance.fromMetres(workout.distanceInMetres.toLong())
 
         name.text = workout.name
         description.text = workout.description
